backend/requirements.txt
flask
langchain
chainlit
python-dotenv
chromadb
torch
transformers
sentence_transformers
unstructured
pypdf
ctransformers
Markdown
langchain_community
llamaapi
langchain_experimental
langchain-google-genai
pillow
psycopg2
langchain_mistralai
langchainhub
sqlalchemy
pyhive
thrift
thrift_sasl
langchain_google_vertexai